backend/requirements.txt
django
django-ninja
psycopg2-binary
python-dotenv
django-cors-headers